test: repartition durable lifecycle regressions
