+++ pkg/services/factory_sessions/internal/execution/fake_service_lifecycle_inspection_test.go
@@ -5,12 +5,7 @@ import (
 	"encoding/json"
 	"errors"
 	"testing"
-	"time"
 
-	"github.com/portpowered/infinite-you/internal/testutil/checkpointfixtures"
-	"github.com/portpowered/infinite-you/internal/testutil/factoryruntimefixtures"
-	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
-	factory "github.com/portpowered/infinite-you/pkg/services/factory_runtime"
 	"github.com/portpowered/infinite-you/pkg/services/recordings"
 )
 
@@ -884,195 +879,4 @@ func (reader *dispatchDurabilityReader) CompletedFlushWatermark(
 ) (recordings.CanonicalEventCursor, bool) {
 	reader.calls++
 	return reader.cursor, reader.available
-}
-
-func TestJavaScriptRuntimeService_CloseRejectsStartWithoutOrphaningReservation(t *testing.T) {
-	t.Parallel()
-
-	t.Run("async", func(t *testing.T) {
-		exerciseStartAdmissionAfterClose(t, func(service *JavaScriptRuntimeService) error {
-			_, err := service.StartAsync(context.Background(), inlineWorkflowStartRequest(
-				"req-runtime-close-admission-async-001",
-				simpleFinalWorkflowSource,
-				map[string]any{"subject": "shutdown"},
-				nil,
-			))
-			return err
-		})
-	})
-
-	t.Run("wait sync", func(t *testing.T) {
-		waitMillis := int64(100)
-		exerciseStartAdmissionAfterClose(t, func(service *JavaScriptRuntimeService) error {
-			request := inlineWorkflowStartRequest(
-				"req-runtime-close-admission-sync-001",
-				simpleFinalWorkflowSource,
-				map[string]any{"subject": "shutdown"},
-				nil,
-			)
-			request.Wait = &WaitOptions{TimeoutMillis: &waitMillis}
-			_, err := service.StartSync(context.Background(), request)
-			return err
-		})
-	})
-}
-
-func exerciseStartAdmissionAfterClose(
-	t *testing.T,
-	start func(*JavaScriptRuntimeService) error,
-) {
-	t.Helper()
-	entered := make(chan struct{})
-	release := make(chan struct{})
-	workflows := &blockingWorkflowDefinitions{
-		JavaScriptWorkflows: scriptedSuccessfulRuntimeWorkflows(map[string]any{"status": "started"}),
-		entered:             entered,
-		release:             release,
-	}
-	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
-		ProjectRoot: t.TempDir(),
-		Workflows:   workflows,
-	})
-	startDone := make(chan error, 1)
-	go func() { startDone <- start(service) }()
-
-	select {
-	case <-entered:
-	case <-time.After(5 * time.Second):
-		t.Fatal("timed out waiting for start preparation")
-	}
-	if err := service.Close(); err != nil {
-		t.Fatalf("Close: %v", err)
-	}
-	close(release)
-	if err := <-startDone; !errors.Is(err, ErrDurableExecutionClosed) {
-		t.Fatalf("start error = %v, want ErrDurableExecutionClosed", err)
-	}
-
-	service.mu.RLock()
-	sessionCount := len(service.sessions)
-	replayCount := len(service.startReplay)
-	inflightCount := len(service.startInflight)
-	service.mu.RUnlock()
-	if sessionCount != 0 || replayCount != 0 || inflightCount != 0 {
-		t.Fatalf(
-			"close-rejected start left durable state: sessions=%d replay=%d inflight=%d",
-			sessionCount, replayCount, inflightCount,
-		)
-	}
-}
-
-type blockingWorkflowDefinitions struct {
-	factory.JavaScriptWorkflows
-	entered chan struct{}
-	release chan struct{}
-}
-
-func (workflows *blockingWorkflowDefinitions) ResolveSource(
-	request factory.WorkflowSourceRequest,
-	context factory.WorkflowSourceContext,
-) factory.WorkflowSourceResolution {
-	close(workflows.entered)
-	<-workflows.release
-	return workflows.JavaScriptWorkflows.ResolveSource(request, context)
-}
-
-func TestJavaScriptRuntimeService_CloseRejectsResumeWithoutOrphaningSession(t *testing.T) {
-	t.Parallel()
-	const sessionID = "dur-sess-0123456789abcdef0123456789abcdef"
-	entered, release := make(chan struct{}), make(chan struct{})
-	checkpointSummaries := &blockingCheckpointSummaries{
-		JavaScriptCheckpointSummaries: checkpointfixtures.CheckpointSummariesFixture{
-			BuildResult:  checkpointfixtures.ResumableCheckpointSummaryResult(),
-			LatestResult: checkpointfixtures.ResumableCheckpointSummaryResult(),
-		},
-		entered: entered, release: release,
-	}
-	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
-		ProjectRoot: t.TempDir(), CheckpointSummaries: checkpointSummaries,
-		Workflows: factoryruntimefixtures.ScriptedJavaScriptWorkflows{},
-	})
-	state := interruptedSessionForAdmissionTest(sessionID)
-	service.mu.Lock()
-	service.sessions[sessionID] = &state
-	service.mu.Unlock()
-
-	resumeDone := make(chan error, 1)
-	go func() {
-		_, err := service.ResumeInterruptedSession(context.Background(), sessionID, ResumeSessionRequest{
-			RequestID: "req-runtime-close-admission-resume-001",
-		})
-		resumeDone <- err
-	}()
-	select {
-	case <-entered:
-	case <-time.After(5 * time.Second):
-		t.Fatal("timed out waiting for resume validation")
-	}
-	if err := service.Close(); err != nil {
-		t.Fatalf("Close: %v", err)
-	}
-	close(release)
-	if err := <-resumeDone; !errors.Is(err, ErrDurableExecutionClosed) {
-		t.Fatalf("resume error = %v, want ErrDurableExecutionClosed", err)
-	}
-	read, err := service.GetSession(context.Background(), sessionID)
-	if err != nil {
-		t.Fatalf("GetSession after rejected resume: %v", err)
-	}
-	if read.Status != LifecycleStatusInterrupted {
-		t.Fatalf("session status after rejected resume = %q, want INTERRUPTED", read.Status)
-	}
-	service.mu.RLock()
-	active := service.sessions[sessionID]
-	var runCancel context.CancelFunc
-	if active != nil {
-		runCancel = active.runCancel
-	}
-	service.mu.RUnlock()
-	if runCancel != nil {
-		t.Fatal("rejected resume left a runnable session cancel function")
-	}
-}
-
-type blockingCheckpointSummaries struct {
-	factory.JavaScriptCheckpointSummaries
-	entered, release chan struct{}
-}
-
-func (summaries *blockingCheckpointSummaries) Latest(
-	input factory.JavaScriptCheckpointSummaryInput,
-) *factory.JavaScriptCheckpointSummary {
-	close(summaries.entered)
-	<-summaries.release
-	return summaries.JavaScriptCheckpointSummaries.Latest(input)
-}
-
-func interruptedSessionForAdmissionTest(sessionID string) runtimeSessionState {
-	startedAt := time.Date(2026, 1, 1, 0, 0, 0, 0, time.UTC)
-	startRequest := StartRequest{RequestID: "req-runtime-close-admission-resume-start-001", Source: Source{
-		Kind: factory.WorkflowSourceKindWorkflowName, WorkflowName: "resumable-two-step-fake-children",
-	}, Args: map[string]any{"subject": "workflows"}}
-	state := runtimeSessionState{
-		session: SessionReadResult{SessionID: sessionID, Status: LifecycleStatusInterrupted,
-			OrchestratorKind: interfaces.OrchestratorKindJavaScript, Dialect: "you-workflow-v1",
-			SourceHash: "sha256:scripted", Lifecycle: &LifecycleTimestamps{StartedAt: &startedAt, InterruptedAt: &startedAt}},
-		result:     ResultReadResult{SessionID: sessionID, SessionStatus: LifecycleStatusInterrupted, ResultStatus: ResultStatusPartial},
-		dispatches: []DispatchSummary{{ID: "dispatch-1", Status: DispatchStatusCompleted, Attempt: 1}},
-		runtimeRecords: []factory.JavaScriptRuntimeRecord{
-			{Sequence: 1, Kind: factory.JavaScriptRecordKindChildDispatch, ChildDispatch: &factory.JavaScriptChildDispatchRecord{
-				DispatchID: "dispatch-1", ChildIndex: 1, Status: factory.JavaScriptChildDispatchStatusCompleted,
-				Output: map[string]any{"text": "step one"},
-			}},
-			{Sequence: 2, Kind: factory.JavaScriptRecordKindCheckpoint, Checkpoint: &factory.JavaScriptCheckpointRecord{
-				ID: "checkpoint-1", Label: "after-step-one",
-			}},
-		},
-		checkpointSummary: checkpointfixtures.ResumableCheckpointSummaryResult(), startRequest: &startRequest,
-		resolvedSource: ResolvedSource{Kind: factory.WorkflowSourceKindWorkflowName,
-			SourceRef: "resumable-two-step-fake-children.workflow.js", SourceHash: "sha256:scripted", Dialect: "you-workflow-v1"},
-		sourceContent: "scripted resumable workflow",
-	}
-	state.events = rebuildRuntimeSessionCanonicalEvents(&state)
-	return state
 }
