+++ pkg/services/factory_sessions/internal/execution/javascript_runtime_result_test.go
@@ -969,43 +969,3 @@ func TestJavaScriptRuntimeService_SnapshotSizeLimitHonorsExactBoundAndRejectsBef
 		t.Fatalf("one-byte-over snapshot = %s, want prior %s", stored, previous)
 	}
 }
-
-func TestJavaScriptRuntimeService_CloseReportsNonCooperativeRunTimeout(t *testing.T) {
-	t.Parallel()
-
-	entered := make(chan struct{})
-	release := make(chan struct{})
-	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
-		ProjectRoot: t.TempDir(),
-		Workflows: scriptedRuntimeWorkflows(func(
-			context.Context,
-			factory.JavaScriptRuntimeRequest,
-			factory.JavaScriptRuntimeHooks,
-		) (factory.JavaScriptRuntimeOutcome, error) {
-			close(entered)
-			<-release
-			return factory.JavaScriptRuntimeOutcome{OK: true}, nil
-		}),
-	})
-	defer func() {
-		close(release)
-		service.runWaitGroup.Wait()
-	}()
-	if _, err := service.StartAsync(context.Background(), inlineWorkflowStartRequest(
-		"req-runtime-close-timeout-001",
-		busyLoopWorkflowSource,
-		nil,
-		nil,
-	)); err != nil {
-		t.Fatalf("StartAsync: %v", err)
-	}
-	select {
-	case <-entered:
-	case <-time.After(time.Second):
-		t.Fatal("timed out waiting for the non-cooperative workflow to start")
-	}
-
-	if err := service.closeWithTimeout(time.Millisecond); !errors.Is(err, ErrDurableExecutionShutdownTimeout) {
-		t.Fatalf("closeWithTimeout error = %v, want ErrDurableExecutionShutdownTimeout", err)
-	}
-}
